requirements.txt
streamlit==0.70.0
pandas==1.1.4
plotly==4.13.0
matplotlib==3.3.2
Pillow==8.0.1
